requirements.txt
Django==5.0.6
django-jazzmin==2.7.0
django-summernote==0.8.20.0
django-taggit==4.0.0
djangorestframework==3.15.1
django-filter==23.5
gunicorn==21.2.0
whitenoise==6.6.0
psycopg2-binary==2.9.9
Pillow==10.2.0
python-decouple==3.8